dashboard api issue fix

See merge request nmr/abdm-nmr!319

+++ src/main/java/in/gov/abdm/nmr/dto/DashboardTO.java
@@ -19,7 +19,7 @@ public class DashboardTO {
     private String councilName;
     private String applicantFullName;
     private BigInteger workFlowStatusId;
-    private Double pendency;
+    private int pendency;
     private String gender;
     private String emailId;
     private String mobileNumber;
